bots/mattbot/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
ddgs>=0.1.0
python-dateutil>=2.8.2
loguru>=0.7.0
